+++ src/Eve-O-Preview/Configuration/Interface/SystemLabelPosition.cs
@@ -1,0 +1,16 @@
+﻿namespace EveOPreview.Configuration
+{
+	public enum SystemLabelPosition
+	{
+		BelowName,
+		NorthWest,
+		North,
+		NorthEast,
+		West,
+		Center,
+		East,
+		SouthWest,
+		South,
+		SouthEast
+	}
+}